TurtleGUI/requirements.txt
colorgram.py==1.2.0
Pillow==9.2.0
